Moved all files to './source'. Added script for copying them to root before npm publish
