+++ .npmignore
@@ -1,10 +1,8 @@
-#tests
-tests
 *.test.*
-
-#other
 .babelrc
 .eslintrc.json
 .gitignore
 .npmrc
+source
+tests
 yarn.lock
