+++ .travis.yml
@@ -5,6 +5,7 @@ install: yarn
 script:
   - yarn build
   - yarn test
+before_deploy: yarn copy
 deploy:
   provider: npm
   email: $NPM_EMAIL
