+++ source/clamp.js
