+++ source/is-equal.js
