+++ source/kebab-to-pascal.js
