+++ source/replace-query-parameters.js
