+++ source/strip-properties-with-value.js
