+++ source/strip-undefined.js
