+++ tests/kebab-to-camel.test.js
@@ -1,6 +1,6 @@
 import test from 'ava';
 
-import kebabToCamel from '../kebab-to-camel';
+import kebabToCamel from '../source/kebab-to-camel';
 
 test('kebab-to-camel', t => {
   t.is(kebabToCamel('kebab-to-camel'), 'kebabToCamel');
