+++ tests/kebab-to-pascal.test.js
@@ -1,6 +1,6 @@
 import test from 'ava';
 
-import kebabToPascal from '../kebab-to-pascal';
+import kebabToPascal from '../source/kebab-to-pascal';
 
 test('kebab-to-pascal', t => {
   t.is(kebabToPascal('kebab-to-pascal'), 'KebabToPascal');
