+++ tests/strip-properties-with-value.test.js
@@ -1,6 +1,6 @@
 import test from 'ava';
 
-import stripPropertiesWithValue from '../strip-properties-with-value';
+import stripPropertiesWithValue from '../source/strip-properties-with-value';
 
 test('strip-properties-with-value', t => {
   t.deepEqual(stripPropertiesWithValue({ a: 1, b: 2 }, 1), { b: 2 });
